% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRycollection-package.R
\docType{package}
\name{pRycollection-package}
\alias{pRycollection}
\alias{pRycollection-package}
\title{pRycollection: Diverse Datasets from Paraguay}
\description{
This dataset contains a variety of datasets from different scientific fields, e.g., crime, healthcare, finance, weather, and others. Some datasets include national data, others cover the five largest Paraguayan cities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schneiderpy/pRycollection/}
  \item Report bugs at \url{https://github.com/schneiderpy/pRycollection/issues/}
}

}
\author{
\strong{Maintainer}: Andreas Schneider \email{schneiderconsultingpy@gmail.com} (\href{https://orcid.org/0000-0001-5630-1097}{ORCID}) [data manager]

}
\keyword{internal}
